Rename common.h to rina-common.h, fix #includes
