+++ components/EFCP/include/delim.h
@@ -9,7 +9,7 @@
 #define COMPONENTS_EFCP_INCLUDE_DELIM_H_
 
 #include "freertos/FreeRTOS.h"
-#include "common.h"
+#include "rina_common.h"
 #include "efcpStructures.h"
 
 typedef struct xDELIM
