+++ components/IPCP/IpcManager.c
@@ -19,7 +19,7 @@
 #include "freertos/semphr.h"
 
 #include "IPCP.h"
-#include "common.h"
+#include "rina_common.h"
 #include "configRINA.h"
 #include "IpcManager.h"
 #include "normalIPCP.h"
