+++ components/IPCP/cepIdm.c
@@ -1,7 +1,7 @@
 
 #include "freertos/FreeRTOS.h"
 
-#include "common.h"
+#include "rina_common.h"
 #include "configRINA.h"
 #include "configSensor.h"
 #include "cepidm.h"
