+++ components/IPCP/include/rina_common.h
