+++ components/IPCP/pidm.c
@@ -7,7 +7,7 @@
  **/
 #include "freertos/FreeRTOS.h"
 
-#include "common.h"
+#include "rina_common.h"
 #include "configRINA.h"
 #include "configSensor.h"
 #include "pidm.h"
